+++ edgy/core/connection/registry.py
@@ -21,6 +21,7 @@ from typing import (
 )
 
 import sqlalchemy
+from loguru import logger
 from sqlalchemy import Engine
 from sqlalchemy.ext.asyncio.engine import AsyncEngine
 from sqlalchemy.orm import declarative_base as sa_declarative_base
@@ -305,22 +306,34 @@ class Registry:
         await database.connect()
         if not self.pattern_models or name in self.dbs_reflected:
             return
-        tmp_metadata = sqlalchemy.MetaData(schema=schema)
-        await database.run_sync(tmp_metadata.reflect)
-        for table in tmp_metadata.tables.values():
-            for pattern_model in self.pattern_models.values():
-                if name not in pattern_model.meta.databases:
-                    continue
-                if pattern_model.meta.pattern.match(table.name):
+        tmp_metadata = sqlalchemy.MetaData()
+        await database.run_sync(tmp_metadata.reflect, schema=schema)
+        try:
+            for table in tmp_metadata.tables.values():
+                for pattern_model in self.pattern_models.values():
+                    if name not in pattern_model.meta.databases:
+                        continue
+                    assert pattern_model.meta.model is pattern_model
+                    # table.key would contain the schema name
+                    if not pattern_model.meta.pattern.match(table.name):
+                        continue
                     new_name = pattern_model.meta.template(table)
-                    old_model = self.get_model(new_name)
+                    old_model: Optional[BaseModelType] = None
+                    with contextlib.suppress(LookupError):
+                        old_model = self.get_model(new_name)
                     if old_model is not None:
                         raise Exception(
                             f"Conflicting model: {old_model.__name__} with pattern model: {pattern_model.__name__}"
                         )
-                    pattern_model = pattern_model.copy_edgy_model(self, name=new_name)
+                    pattern_model = pattern_model.copy_edgy_model(name=new_name)
                     pattern_model.meta.tablename = table.name
-        self.dbs_reflected.add(name)
+                    pattern_model.meta.concrete = True
+                    pattern_model.add_to_registry(self)
+
+            self.dbs_reflected.add(name)
+        except BaseException as exc:
+            await database.disconnect()
+            raise exc
 
     async def __aenter__(self) -> "Registry":
         dbs: List[Tuple[Union[str, None], Database]] = [(None, self.database)]
@@ -335,6 +348,8 @@ class Registry:
             for num, value in enumerate(results):
                 if not isinstance(value, BaseException):
                     ops2.append(dbs[num][1].disconnect())
+                else:
+                    logger.opt(exception=value).error("Failed to connect database.")
             await asyncio.gather(*ops2)
         return self